fixed failed github actions
