+++ tests/BaseTestCase.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Tests\Kilip\Laravel\Alice;
 
-use Doctrine\Persistence\ObjectManager;
 use Kilip\Laravel\Alice\AliceServiceProvider;
 use LaravelDoctrine\ORM\DoctrineServiceProvider;
 use Orchestra\Testbench\TestCase;
@@ -38,22 +37,5 @@ abstract class BaseTestCase extends TestCase
         ]);
 
         $config->set('alice.loader', 'alice.loader.doctrine');
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return ObjectManager
-     */
-    protected function getEntityManager($name = null)
-    {
-        /** @var \Doctrine\Persistence\ManagerRegistry $registry */
-        $registry = $this->app->get('registry');
-
-        if (null === $name) {
-            $name = $registry->getDefaultManagerName();
-        }
-
-        return $registry->getManager($name);
     }
 }
