+++ tests/Resources/eloquent-migrations/2020_07_14_095300_create_eloquent_groups_table.php
@@ -15,7 +15,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateGroupsTable extends Migration
+class CreateEloquentGroupsTable extends Migration
 {
     public const TABLE_NAME = 'eloquent_groups';
 
@@ -27,7 +27,7 @@ class CreateGroupsTable extends Migration
     public function up()
     {
         Schema::create(static::TABLE_NAME, function (Blueprint $table) {
-            $table->id();
+            $table->integer('id')->primary();
             $table->string('name');
             $table->timestamps();
         });
